making componentStatusNotify function accessible via ws handler in front-end.
* moved notifyComponentStatus to util
